fix(ios): fix vehicle decoding error - decimal field type mismatches
totalFuelCost and totalMaintenanceCost were String? but API returns
numbers. currentHours was Int? but model is now DecimalField. hours in
LogEventRequest was Int but needs Double. Bump build to 6.

+++ Sources/HaleHubIOS/Features/Vehicles/LogEventSheet.swift
@@ -147,7 +147,7 @@ struct LogEventSheet: View {
 
         var body = LogEventRequest(eventType: eventType, date: formatter.string(from: date))
         body.miles = vehicle.isBoat ? nil : Int(odometer)
-        body.hours = vehicle.isBoat ? Int(odometer) : nil
+        body.hours = vehicle.isBoat ? Double(odometer) : nil
         body.gallons = Double(gallons)
         body.pricePerGallon = Double(pricePerGallon)
         body.notes = notes.isEmpty ? nil : notes
